modified calculations

+++ cpe/src/org/cougaar/cpe/agents/plugin/ControlUnitAgentPlugin.java
@@ -67,7 +67,7 @@ public class ControlUnitAgentPlugin extends ComponentPlugin {
 	 */
 	public double[][] getMeanDelay(long currentTime, int period) {
 		double[][] avg_delays = new double[7][2];
-		long[][] total_delay = new long[7][3];
+		double[][] total_delay = new double[7][3];
 		Iterator iter = null;
 		int count = 0;
 
@@ -139,7 +139,7 @@ public class ControlUnitAgentPlugin extends ComponentPlugin {
 						if (total_delay[ii][1] > 0) {
 
 							avg_delays[ii][0] =
-								total_delay[ii][0] / total_delay[ii][1];
+								(double)(total_delay[ii][0] / total_delay[ii][1]);
 						} else {
 
 							avg_delays[ii][0] = -1;
